requirements.txt
Flask
flask-cors
pillow
pytesseract
